core/gpu_sw: drop redundant uint16_t intermediates in 5551/565 fast paths
CopyOutRow16<RGBA5551> and CopyOutRow16<RGB565> use SSE2's
_mm_set1_epi16(int16_t) to build immediate AND/SLLI masks
(0x3E0 and 0x20). Each callsite went:

  static_cast<int16_t>(static_cast<uint16_t>(0x3E0))

The uint16_t intermediate was guarding against narrowing from
int (which is what an unadorned 0x3E0 is), but 0x3E0 (992)
and 0x20 (32) both fit comfortably in int16_t's positive range
(max 32767), so static_cast<int16_t>(0x3E0) is well-defined on
every supported compiler without the intermediate. Drop the
extra cast at all three sites.

Codegen-identical (verified by MD5 of the .so before and after
the change). 3 sites, 0 lines net.

+++ src/core/gpu_sw.cpp
@@ -149,7 +149,7 @@ ALWAYS_INLINE void CopyOutRow16<HostDisplayPixelFormat::RGBA5551, uint16_t>(cons
     const __m128i single_mask = _mm_set1_epi16(0x1F);
     __m128i value = _mm_loadu_si128(reinterpret_cast<const __m128i*>(src_ptr));
     src_ptr += 8;
-    __m128i a = _mm_and_si128(value, _mm_set1_epi16(static_cast<int16_t>(static_cast<uint16_t>(0x3E0))));
+    __m128i a = _mm_and_si128(value, _mm_set1_epi16(static_cast<int16_t>(0x3E0)));
     __m128i b = _mm_and_si128(_mm_srli_epi16(value, 10), single_mask);
     __m128i c = _mm_slli_epi16(_mm_and_si128(value, single_mask), 10);
     value = _mm_or_si128(_mm_or_si128(a, b), c);
@@ -188,8 +188,8 @@ ALWAYS_INLINE void CopyOutRow16<HostDisplayPixelFormat::RGB565, uint16_t>(const 
     const __m128i single_mask = _mm_set1_epi16(0x1F);
     __m128i value = _mm_loadu_si128(reinterpret_cast<const __m128i*>(src_ptr));
     src_ptr += 8;
-    __m128i a = _mm_slli_epi16(_mm_and_si128(value, _mm_set1_epi16(static_cast<int16_t>(static_cast<uint16_t>(0x3E0)))), 1);
-    __m128i b = _mm_slli_epi16(_mm_and_si128(value, _mm_set1_epi16(static_cast<int16_t>(static_cast<uint16_t>(0x20)))), 1);
+    __m128i a = _mm_slli_epi16(_mm_and_si128(value, _mm_set1_epi16(static_cast<int16_t>(0x3E0))), 1);
+    __m128i b = _mm_slli_epi16(_mm_and_si128(value, _mm_set1_epi16(static_cast<int16_t>(0x20))), 1);
     __m128i c = _mm_and_si128(_mm_srli_epi16(value, 10), single_mask);
     __m128i d = _mm_slli_epi16(_mm_and_si128(value, single_mask), 11);
     value = _mm_or_si128(_mm_or_si128(_mm_or_si128(a, b), c), d);
